move test case to seperate file
