+++ package.json
@@ -5,8 +5,8 @@
   "main": "app.js",
   "scripts": {
     "wp": "webpack-dev-server --hot --history-api-fallback --open",
-    "test": "cross-env NODE_ENV=test nyc mocha server/test/test.js ",
-    "test:db": "set NODE_ENV=test:s&&sequelize db:migrate:undo:all&&sequelize db:migrate  && cross-env NODE_ENV=test nyc mocha server/test/test.js",
+    "test": "cross-env NODE_ENV=test nyc mocha server/test/ ",
+    "test:db": "set NODE_ENV=test:s&&sequelize db:migrate:undo:all&&sequelize db:migrate  && cross-env NODE_ENV=test nyc mocha server/test/",
     "s:test": "set NODE_ENV=test:s&& mocha --compilers js:babel-core/register --timeout 10000 server/test/test.js",
     "heroku-postbuild":"sequelize db:migrate",
     "migrate": "sequelize db:migrate:undo:all&&sequelize db:migrate",
